requirements.txt
python-telegram-bot
python-dotenv
yt-dlp
mutagen
Pillow
requests
pydub